feat: 응답 Dto 추가, respository 구조 통일

+++ backend/src/main/java/com/team3/airdnd/payment/domain/Payment.java
@@ -24,7 +24,7 @@ public class Payment {
     @JoinColumn(name = "payment_method_id", nullable = false)
     private PaymentMethod paymentMethod;
 
-    private Integer amount;
+    private Long amount;
 
     @Column(name = "paid_at", nullable = false)
     private LocalDateTime paidAt;
